fix: adjust existing icons
