--- packages/ocean-icons-react/src/BluLogo.tsx
@@ -1,30 +1,0 @@
-import React from 'react';
-
-type IconProps = {
-  size?: number;
-} & React.SVGProps<SVGSVGElement>;
-
-const BluLogo = (
-  { size = 24, ...props }: IconProps,
-  svgRef: React.Ref<SVGSVGElement>
-) => (
-  <svg
-    width={size}
-    height={size}
-    viewBox="0 0 24 24"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-    ref={svgRef}
-    {...props}
-  >
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M7.61538 2C6.17095 2 5 3.17094 5 4.61538V19.3846C5 20.8291 6.17095 22 7.61538 22H15C17.9738 22 20.3846 19.5892 20.3846 16.6154V15.8462C20.3846 13.9447 19.399 12.2735 17.9109 11.3154C18.9599 10.3329 19.6154 8.93529 19.6154 7.38462C19.6154 4.41077 17.2046 2 14.2308 2H7.61538ZM7.61538 19.6923C7.44545 19.6923 7.30769 19.5546 7.30769 19.3846V4.61538C7.30769 4.44545 7.44545 4.30769 7.61538 4.30769H14.2308C15.9301 4.30769 17.3077 5.68528 17.3077 7.38462C17.3077 8.52351 16.6889 9.51789 15.7692 10.0499V12.8662C17.0963 13.2077 18.0769 14.4124 18.0769 15.8462V16.6154C18.0769 18.3147 16.6993 19.6923 15 19.6923H7.61538ZM13.4615 8.15385C13.4615 8.72329 13.1522 9.22048 12.6923 9.48649V14.5135C13.1522 14.7795 13.4615 15.2767 13.4615 15.8462C13.4615 16.6958 12.7727 17.3846 11.9231 17.3846C11.0734 17.3846 10.3846 16.6958 10.3846 15.8462C10.3846 15.2767 10.694 14.7795 11.1538 14.5135V9.48649C10.694 9.22048 10.3846 8.72329 10.3846 8.15385C10.3846 7.30418 11.0734 6.61538 11.9231 6.61538C12.7727 6.61538 13.4615 7.30418 13.4615 8.15385Z"
-      fill="currentColor"
-    />
-  </svg>
-);
-
-const ForwardRef = React.forwardRef(BluLogo);
-export default ForwardRef;
