+++ packages/ocean-icons-react/src/Pix.tsx
@@ -11,17 +11,15 @@ const Pix = (
   <svg
     xmlns="http://www.w3.org/2000/svg"
     viewBox="0 0 24 24"
-    fill="none"
+    fill="currentColor"
     width={size}
     height={size}
     ref={svgRef}
     {...props}
   >
     <path
-      fillRule="evenodd"
+      d="M2.559 13.349l1.196 1.196.586.518.021.02.021.02c.538.535 1.388.897 2.117.897.656 0 1.25-.3 1.848-.893.103-.105.212-.22.323-.339.34-.36.712-.754 1.095-1.043C10.365 13.273 11.06 13 12 13c.963 0 1.651.305 2.234.761.267.21.507.446.72.662l.115.116c.18.183.352.356.548.537l.015.013.014.014c.6.595 1.196.897 1.854.897.729 0 1.577-.362 2.115-.897l.021-.02.022-.02.596-.527 1.187-1.187a1.908 1.908 0 000-2.698l-1.194-1.194-.588-.52-.022-.02-.02-.02C19.079 8.362 18.229 8 17.5 8c-.656 0-1.25.3-1.848.893-.104.106-.212.22-.324.339-.339.36-.71.754-1.094 1.043-.599.452-1.294.725-2.234.725-.963 0-1.651-.305-2.234-.76a7.804 7.804 0 01-.72-.663l-.116-.116c-.18-.182-.351-.356-.548-.536l-.014-.014-.015-.014C7.755 8.302 7.157 8 6.5 8c-.729 0-1.577.362-2.116.897l-.02.02-.022.02-.596.527L2.56 10.65a1.908 1.908 0 000 2.698zm4.598-7.296c1.046.169 1.903.73 2.592 1.411.23.212.437.422.614.602l.103.103c.211.214.374.37.533.495.27.212.533.336 1 .336.492 0 .767-.123 1.031-.322.24-.18.438-.39.722-.692.135-.144.29-.308.477-.499l.009-.008c.691-.688 1.552-1.256 2.605-1.426l-3.494-3.494a1.908 1.908 0 00-2.698 0L7.157 6.053zm3.494 15.388l-3.494-3.493c1.053-.171 1.914-.739 2.605-1.427l.009-.008c.187-.191.342-.355.477-.499.284-.301.483-.512.722-.692.264-.2.54-.322 1.03-.322.468 0 .73.124 1.001.336.16.124.322.281.533.495l.102.103c.178.18.385.39.615.602.69.681 1.546 1.242 2.592 1.412l-3.494 3.493a1.908 1.908 0 01-2.698 0z"
       clipRule="evenodd"
-      d="M2.55867 13.3488L3.75516 14.5452L4.34076 15.0634L4.36249 15.0827L4.38307 15.1031C4.92116 15.638 5.77097 16 6.49993 16C7.15556 16 7.74996 15.7002 8.34772 15.1074C8.4515 15.0015 8.55969 14.8868 8.67148 14.7682C9.01052 14.4085 9.38272 14.0137 9.76551 13.725C10.3647 13.2733 11.06 13 11.9999 13C12.9629 13 13.6512 13.3052 14.2336 13.761C14.5012 13.9704 14.741 14.2073 14.9547 14.4228L15.0695 14.5389L15.0695 14.5389L15.0695 14.5389C15.2497 14.7215 15.4213 14.8954 15.6175 15.0755L15.6323 15.089L15.6465 15.1031C16.2453 15.6984 16.8423 16 17.4999 16C18.2285 16 19.0771 15.6383 19.6154 15.1031L19.6361 15.0826L19.6579 15.0633L20.2543 14.5358L21.4413 13.3488C22.1862 12.6039 22.1862 11.3961 21.4413 10.6512L20.2469 9.45681L19.6591 8.93665L19.6373 8.91742L19.6168 8.89697C19.0787 8.36207 18.2289 8.00005 17.4999 8.00005C16.8443 8.00005 16.2499 8.29986 15.6521 8.89268C15.5483 8.99856 15.4401 9.11333 15.3284 9.23192C14.9893 9.59159 14.6171 9.98644 14.2343 10.2751C13.6351 10.7268 12.9398 11.0001 11.9999 11.0001C11.037 11.0001 10.3487 10.6949 9.7662 10.2391C9.4986 10.0296 9.25883 9.79283 9.04516 9.57728L8.93038 9.46115C8.75019 9.27857 8.57857 9.10467 8.38232 8.92463L8.36756 8.91109L8.35336 8.89697C7.75451 8.30167 7.15749 8.00005 6.4999 8.00005C5.77129 8.00005 4.92273 8.36181 4.38439 8.89697L4.36373 8.9175L4.34192 8.93679L3.74622 9.4637L2.55868 10.6512C1.81378 11.3961 1.81377 12.6039 2.55867 13.3488ZM7.15738 6.05253C8.20315 6.22193 9.05957 6.78297 9.74893 7.46426C9.97938 7.67649 10.1862 7.8862 10.3633 8.06579L10.3634 8.06583L10.4656 8.16928C10.6774 8.38295 10.8398 8.53959 10.9988 8.66403C11.2696 8.87594 11.5323 9.00005 11.9999 9.00005C12.4906 9.00005 12.766 8.87737 13.0303 8.67811C13.2693 8.4979 13.4677 8.28739 13.7519 7.9858C13.887 7.84241 14.0415 7.67844 14.229 7.48735L14.2377 7.4785L14.2378 7.47855C14.9294 6.79109 15.7895 6.22313 16.8426 6.05252L13.3488 2.55868C12.6039 1.81377 11.3961 1.81377 10.6512 2.55868L7.15738 6.05253ZM10.6512 21.4413L7.15745 17.9475C8.21043 17.7769 9.07052 17.209 9.76204 16.5215L9.76209 16.5216L9.77079 16.5127C9.9583 16.3216 10.1128 16.1577 10.248 16.0143L10.248 16.0143C10.5322 15.7127 10.7306 15.5022 10.9696 15.322C11.2338 15.1227 11.5093 15 11.9999 15C12.4676 15 12.7303 15.1242 13.001 15.3361C13.1601 15.4605 13.3225 15.6171 13.5343 15.8308L13.6365 15.9343L13.6365 15.9343L13.6365 15.9343C13.8136 16.1139 14.0205 16.3236 14.2509 16.5358C14.9403 17.2171 15.7967 17.7782 16.8425 17.9476L13.3488 21.4413C12.6039 22.1862 11.3961 22.1862 10.6512 21.4413Z"
-      fill="#0025E0"
     />
   </svg>
 );
