+++ packages/ocean-icons-react/src/PresentationQrcodeOutline.tsx
@@ -12,12 +12,13 @@ const PresentationQrcodeOutline = (
     xmlns="http://www.w3.org/2000/svg"
     fill="currentColor"
     viewBox="0 0 24 25"
+    stroke="currentColor"
     width={size}
     height={size}
     ref={svgRef}
     {...props}
   >
-    <path d="M21 22.5C21.5523 22.5 22 22.0523 22 21.5C22 20.9477 21.5523 20.5 21 20.5V22.5ZM20 21.5V22.5V21.5ZM3 20.5C2.44772 20.5 2 20.9477 2 21.5C2 22.0523 2.44772 22.5 3 22.5V20.5ZM4 4.73077L3 4.73077L4 4.73077ZM5 3.5V2.5V3.5ZM19 3.5V4.5V3.5ZM20 4.73077L21 4.73077L20 4.73077ZM9 16.5L9 17.5H9V16.5ZM10 17.5C10.5523 17.5 11 17.0523 11 16.5C11 15.9477 10.5523 15.5 10 15.5V17.5ZM8 15.5H7H8ZM9 14.5C9 13.9477 8.55229 13.5 8 13.5C7.44772 13.5 7 13.9477 7 14.5H9ZM15 8.5L15 7.5H15V8.5ZM14 7.5C13.4477 7.5 13 7.94771 13 8.5C13 9.05228 13.4477 9.5 14 9.5V7.5ZM16 9.5H15H16ZM15 10.5C15 11.0523 15.4477 11.5 16 11.5C16.5523 11.5 17 11.0523 17 10.5H15ZM21 20.5H20V22.5H21V20.5ZM20 20.5L4 20.5L4 22.5L20 22.5V20.5ZM4 20.5H3V22.5H4V20.5ZM5 21.5L5 4.73077L3 4.73077L3 21.5H5ZM5 4.73077C5 4.6101 5.04069 4.52592 5.06901 4.49107L3.51678 3.22989C3.17002 3.65667 3 4.19859 3 4.73077L5 4.73077ZM5.06901 4.49107C5.08115 4.47613 5.08428 4.47808 5.0715 4.48459C5.05782 4.49157 5.03273 4.5 5 4.5V2.5C4.39463 2.5 3.86814 2.79744 3.51678 3.22989L5.06901 4.49107ZM5 4.5L19 4.5V2.5L5 2.5V4.5ZM19 4.5C18.9673 4.5 18.9422 4.49157 18.9285 4.4846C18.9157 4.47808 18.9189 4.47613 18.931 4.49108L20.4832 3.22989C20.1319 2.79744 19.6054 2.5 19 2.5V4.5ZM18.931 4.49108C18.9593 4.52592 19 4.61011 19 4.73077L21 4.73077C21 4.19859 20.83 3.65667 20.4832 3.22989L18.931 4.49108ZM19 4.73077L19 21.5H21L21 4.73077L19 4.73077ZM14 14.5H15V12.5H14V14.5ZM15 14.5V15.5H17V14.5H15ZM15 15.5H14V17.5H15V15.5ZM14 15.5V14.5H12V15.5H14ZM14 15.5H12C12 16.6046 12.8954 17.5 14 17.5V15.5ZM15 15.5V15.5V17.5C16.1046 17.5 17 16.6046 17 15.5H15ZM15 14.5H17C17 13.3954 16.1046 12.5 15 12.5V14.5ZM14 12.5C12.8954 12.5 12 13.3954 12 14.5H14V12.5ZM9 9.5H10V7.5H9V9.5ZM10 9.5V10.5H12V9.5H10ZM10 10.5H9V12.5H10V10.5ZM9 10.5V9.5H7V10.5H9ZM9 10.5H7C7 11.6046 7.89543 12.5 9 12.5V10.5ZM10 10.5V12.5C11.1046 12.5 12 11.6046 12 10.5H10ZM10 9.5H12C12 8.39543 11.1046 7.5 10 7.5V9.5ZM9 7.5C7.89543 7.5 7 8.39543 7 9.5H9V7.5ZM9 17.5H10V15.5H9V17.5ZM9 15.5V14.5H7V15.5H9ZM9 15.5L9 15.5H7C7 16.6046 7.89543 17.5 9 17.5L9 15.5ZM15 7.5H14V9.5H15V7.5ZM15 9.5V10.5H17V9.5H15ZM15 9.5L15 9.5H17C17 8.39543 16.1046 7.5 15 7.5L15 9.5Z" />
+    <path d="M21 22.5a1 1 0 100-2v2zm-1-1v1-1zm-17-1a1 1 0 100 2v-2zM4 4.73H3h1zM5 3.5v-1 1zm14 0v1-1zm1 1.23h1-1zM9 16.5v1-1zm1 1a1 1 0 100-2v2zm-2-2H7h1zm1-1a1 1 0 10-2 0h2zm6-6v-1 1zm-1-1a1 1 0 100 2v-2zm2 2h-1 1zm-1 1a1 1 0 102 0h-2zm6 10h-1v2h1v-2zm-1 0H4v2h16v-2zm-16 0H3v2h1v-2zm1 1V4.73H3V21.5h2zM5 4.73c0-.12.04-.204.069-.239L3.517 3.23A2.387 2.387 0 003 4.73h2zm.069-.239c.012-.015.015-.013.003-.006A.162.162 0 015 4.5v-2c-.605 0-1.132.297-1.483.73l1.552 1.26zM5 4.5h14v-2H5v2zm14 0a.162.162 0 01-.072-.015c-.012-.007-.01-.009.003.006l1.552-1.261A1.912 1.912 0 0019 2.5v2zm-.069-.009a.397.397 0 01.069.24h2c0-.532-.17-1.074-.517-1.501l-1.552 1.26zm.069.24V21.5h2V4.73h-2zM14 14.5h1v-2h-1v2zm1 0v1h2v-1h-2zm0 1h-1v2h1v-2zm-1 0v-1h-2v1h2zm0 0h-2a2 2 0 002 2v-2zm1 0v2a2 2 0 002-2h-2zm0-1h2a2 2 0 00-2-2v2zm-1-2a2 2 0 00-2 2h2v-2zm-5-3h1v-2H9v2zm1 0v1h2v-1h-2zm0 1H9v2h1v-2zm-1 0v-1H7v1h2zm0 0H7a2 2 0 002 2v-2zm1 0v2a2 2 0 002-2h-2zm0-1h2a2 2 0 00-2-2v2zm-1-2a2 2 0 00-2 2h2v-2zm0 10h1v-2H9v2zm0-2v-1H7v1h2zm0 0H7a2 2 0 002 2v-2zm6-8h-1v2h1v-2zm0 2v1h2v-1h-2zm0 0h2a2 2 0 00-2-2v2z" />
   </svg>
 );
