+++ packages/ocean-icons-react/src/SupplierOutline.tsx
@@ -10,15 +10,15 @@ const SupplierOutline = (
 ) => (
   <svg
     xmlns="http://www.w3.org/2000/svg"
-    fill="currentColor"
+    fill="none"
     viewBox="0 0 24 24"
-    stroke="none"
+    stroke="currentColor"
     width={size}
     height={size}
     ref={svgRef}
     {...props}
   >
-    <path d="M3.4 6.11351L4.16757 6.75448L4.16757 6.75448L3.4 6.11351ZM5.4 3.7185L6.16757 4.35947L6.16757 4.35947L5.4 3.7185ZM18.6 3.7185L19.3676 3.07753L19.3676 3.07753L18.6 3.7185ZM20.6 6.11351L21.3676 5.47254L21.3676 5.47254L20.6 6.11351ZM11 6.69231C11 7.24459 11.4477 7.69231 12 7.69231C12.5523 7.69231 13 7.24459 13 6.69231H11ZM9 10C8.44772 10 8 10.4477 8 11C8 11.5523 8.44772 12 9 12V10ZM15 12C15.5523 12 16 11.5523 16 11C16 10.4477 15.5523 10 15 10V12ZM2 7.1889V19.2066H4V7.1889H2ZM5 22H19V20H5V22ZM22 19.2037V7.19127H20V19.2037H22ZM4.16757 6.75448L6.16757 4.35947L4.63243 3.07753L2.63243 5.47254L4.16757 6.75448ZM17.8324 4.35947L19.8324 6.75448L21.3676 5.47254L19.3676 3.07753L17.8324 4.35947ZM17 4C17.3562 4 17.6591 4.15185 17.8324 4.35947L19.3676 3.07753C18.7855 2.38054 17.9028 2 17 2V4ZM6.16757 4.35947C6.34095 4.15185 6.6438 4 7 4V2C6.09717 2 5.21447 2.38054 4.63243 3.07753L6.16757 4.35947ZM22 7.19127C22 6.55779 21.7705 5.95505 21.3676 5.47254L19.8324 6.75448C19.9488 6.89383 20 7.04743 20 7.19127H22ZM19 22C20.5531 22 22 20.8464 22 19.2037H20C20 19.5452 19.656 20 19 20V22ZM2 19.2066C2 20.8512 3.44906 22 5 22V20C4.3418 20 4 19.5461 4 19.2066H2ZM4 7.1889C4 7.04654 4.05021 6.89501 4.16757 6.75448L2.63243 5.47254C2.2305 5.95386 2 6.55394 2 7.1889H4ZM13 6.69231V3H11V6.69231H13ZM7 4H12V2H7V4ZM12 4H17V2H12V4ZM3 8H21V6H3V8ZM9 12H15V10H9V12Z" />
+    <path d="M3.4 6.114l.768.64-.768-.64zm2-2.395l.768.64-.768-.64zm13.2 0l.768-.641-.768.64zm2 2.395l.768-.641-.768.64zm-9.6.578a1 1 0 102 0h-2zM9 10a1 1 0 100 2v-2zm6 2a1 1 0 100-2v2zM2 7.189v12.018h2V7.189H2zM5 22h14v-2H5v2zm17-2.796V7.19h-2v12.013h2zM4.168 6.754l2-2.395-1.536-1.281-2 2.395 1.536 1.281zM17.832 4.36l2 2.395 1.536-1.281-2-2.395-1.536 1.281zM17 4c.356 0 .66.152.832.36l1.536-1.282C18.786 2.38 17.903 2 17 2v2zm-10.832.36C6.34 4.151 6.644 4 7 4V2c-.903 0-1.786.38-2.368 1.078l1.536 1.281zM22 7.19a2.68 2.68 0 00-.632-1.718l-1.536 1.281A.68.68 0 0120 7.19h2zM19 22c1.553 0 3-1.154 3-2.796h-2c0 .341-.344.796-1 .796v2zM2 19.207C2 20.85 3.45 22 5 22v-2c-.658 0-1-.454-1-.793H2zM4 7.189c0-.142.05-.294.168-.435L2.632 5.473A2.673 2.673 0 002 7.189h2zm9-.497V3h-2v3.692h2zM7 4h5V2H7v2zm5 0h5V2h-5v2zM3 8h18V6H3v2zm6 4h6v-2H9v2z" />
   </svg>
 );
