+++ packages/ocean-icons-react/src/VariableTaxOutline.tsx
@@ -19,19 +19,19 @@ const VariableTaxOutline = (
     {...props}
   >
     <path
-      d="M4 8.62719H3V6L4 8.62719ZM4 8.62719C4.74683 6.78006 5.74014 5.23292 7.46066 4.22823C9.18117 3.22353 11.1864 2.81808 13.1621 3.0754C15.1378 3.33272 16.9723 4.23828 18.378 5.65015C19.7838 7.06202 20.6815 8.90042 20.9303 10.8772M4 8.62719H6M20 15.3773H21V18L20 15.3773ZM20 15.3773C19.2521 17.2233 18.2595 18.7693 16.5392 19.773C14.8188 20.7768 12.8142 21.1818 10.8391 20.9246C8.86402 20.6674 7.02997 19.7624 5.6241 18.3515C4.21823 16.9406 3.3199 15.1033 3.06975 13.1273M20 15.3773H18"
-      strokeWidth="2"
       strokeLinecap="round"
       strokeLinejoin="round"
+      strokeWidth={2}
+      d="M4 8.627H3V6l1 2.627zm0 0c.747-1.847 1.74-3.394 3.46-4.399a9.001 9.001 0 0113.47 6.65M4 8.627h2m14 6.75h1V18l-1-2.623zm0 0c-.748 1.845-1.74 3.391-3.46 4.395a9.004 9.004 0 01-13.47-6.646M20 15.377h-2"
     />
     <path
-      d="M9 15L15 9M10 9.5C10 9.63261 9.94732 9.75979 9.85355 9.85355C9.75979 9.94732 9.63261 10 9.5 10C9.36739 10 9.24021 9.94732 9.14645 9.85355C9.05268 9.75979 9 9.63261 9 9.5C9 9.36739 9.05268 9.24021 9.14645 9.14645C9.24021 9.05268 9.36739 9 9.5 9C9.63261 9 9.75979 9.05268 9.85355 9.14645C9.94732 9.24021 10 9.36739 10 9.5ZM15 14.5C15 14.6326 14.9473 14.7598 14.8536 14.8536C14.7598 14.9473 14.6326 15 14.5 15C14.3674 15 14.2402 14.9473 14.1464 14.8536C14.0527 14.7598 14 14.6326 14 14.5C14 14.3674 14.0527 14.2402 14.1464 14.1464C14.2402 14.0527 14.3674 14 14.5 14C14.6326 14 14.7598 14.0527 14.8536 14.1464C14.9473 14.2402 15 14.3674 15 14.5Z"
-      strokeWidth="2"
       strokeLinecap="round"
       strokeLinejoin="round"
+      strokeWidth={2}
+      d="M9 15l6-6m-5 .5a.5.5 0 11-1 0 .5.5 0 011 0zm5 5a.5.5 0 11-1 0 .5.5 0 011 0z"
     />
-    <circle cx="14.5" cy="14.5" r="1.5" />
-    <circle cx="9.5" cy="9.5" r="1.5" />
+    <circle cx={14.5} cy={14.5} r={1.5} />
+    <circle cx={9.5} cy={9.5} r={1.5} />
   </svg>
 );
