+++ packages/ocean-icons-react/src/index.ts
@@ -163,6 +163,7 @@ export { default as PhoneMissedCall } from './PhoneMissedCall';
 export { default as PhoneOutgoing } from './PhoneOutgoing';
 export { default as Phone } from './Phone';
 export { default as Photograph } from './Photograph';
+export { default as Pix } from './Pix';
 export { default as Placeholder } from './Placeholder';
 export { default as Play } from './Play';
 export { default as PlusCircle } from './PlusCircle';
@@ -269,6 +270,7 @@ export { default as BanOutline } from './BanOutline';
 export { default as BarcodeOutline } from './BarcodeOutline';
 export { default as BeakerOutline } from './BeakerOutline';
 export { default as BellOutline } from './BellOutline';
+export { default as BluLogoOutline } from './BluLogoOutline';
 export { default as BookOpenOutline } from './BookOpenOutline';
 export { default as BookmarkAltOutline } from './BookmarkAltOutline';
 export { default as BookmarkOutline } from './BookmarkOutline';
@@ -342,7 +344,7 @@ export { default as ExclamationOutline } from './ExclamationOutline';
 export { default as ExternalLinkOutline } from './ExternalLinkOutline';
 export { default as EyeOffOutline } from './EyeOffOutline';
 export { default as EyeOutline } from './EyeOutline';
-export { default as SupplierOutline } from './SupplierOutline';
+export { default as FactoryOutline } from './FactoryOutline';
 export { default as FastForwardOutline } from './FastForwardOutline';
 export { default as FilmOutline } from './FilmOutline';
 export { default as FilterOutline } from './FilterOutline';
@@ -378,6 +380,7 @@ export { default as LinkOutline } from './LinkOutline';
 export { default as LocationMarkerOutline } from './LocationMarkerOutline';
 export { default as LockClosedOutline } from './LockClosedOutline';
 export { default as LockOpenOutline } from './LockOpenOutline';
+export { default as LockTaxOutline } from './LockTaxOutline';
 export { default as LoginOutline } from './LoginOutline';
 export { default as LogoutOutline } from './LogoutOutline';
 export { default as MailOpenOutline } from './MailOpenOutline';
@@ -398,10 +401,10 @@ export { default as MusicNoteOutline } from './MusicNoteOutline';
 export { default as NewspaperOutline } from './NewspaperOutline';
 export { default as OfficeBuildingOutline } from './OfficeBuildingOutline';
 export { default as OrderOutline } from './OrderOutline';
-export { default as PaymentOutline } from './PaymentOutline';
 export { default as PaperAirplaneOutline } from './PaperAirplaneOutline';
 export { default as PaperClipOutline } from './PaperClipOutline';
 export { default as PauseOutline } from './PauseOutline';
+export { default as PaymentOutline } from './PaymentOutline';
 export { default as PencilAltOutline } from './PencilAltOutline';
 export { default as PencilOutline } from './PencilOutline';
 export { default as PhoneIncomingOutline } from './PhoneIncomingOutline';
@@ -409,6 +412,7 @@ export { default as PhoneMissedCallOutline } from './PhoneMissedCallOutline';
 export { default as PhoneOutgoingOutline } from './PhoneOutgoingOutline';
 export { default as PhoneOutline } from './PhoneOutline';
 export { default as PhotographOutline } from './PhotographOutline';
+export { default as PixOutline } from './PixOutline';
 export { default as PlaceholderOutline } from './PlaceholderOutline';
 export { default as PlayOutline } from './PlayOutline';
 export { default as PlusCircleOutline } from './PlusCircleOutline';
@@ -416,6 +420,7 @@ export { default as PlusSmOutline } from './PlusSmOutline';
 export { default as PlusOutline } from './PlusOutline';
 export { default as PresentationChartBarOutline } from './PresentationChartBarOutline';
 export { default as PresentationChartLineOutline } from './PresentationChartLineOutline';
+export { default as PresentationQrcodeOutline } from './PresentationQrcodeOutline';
 export { default as PrinterOutline } from './PrinterOutline';
 export { default as PuzzleOutline } from './PuzzleOutline';
 export { default as QrcodeOutline } from './QrcodeOutline';
@@ -450,6 +455,7 @@ export { default as StatusOfflineOutline } from './StatusOfflineOutline';
 export { default as StatusOnlineOutline } from './StatusOnlineOutline';
 export { default as StopOutline } from './StopOutline';
 export { default as SunOutline } from './SunOutline';
+export { default as SupplierOutline } from './SupplierOutline';
 export { default as SupportOutline } from './SupportOutline';
 export { default as SwitchHorizontalOutline } from './SwitchHorizontalOutline';
 export { default as SwitchVerticalOutline } from './SwitchVerticalOutline';
@@ -472,6 +478,7 @@ export { default as UserGroupOutline } from './UserGroupOutline';
 export { default as UserRemoveOutline } from './UserRemoveOutline';
 export { default as UserOutline } from './UserOutline';
 export { default as UsersOutline } from './UsersOutline';
+export { default as VariableTaxOutline } from './VariableTaxOutline';
 export { default as VariableOutline } from './VariableOutline';
 export { default as VideoCameraOutline } from './VideoCameraOutline';
 export { default as ViewBoardsOutline } from './ViewBoardsOutline';
@@ -486,9 +493,3 @@ export { default as XOutline } from './XOutline';
 export { default as ZeroOutline } from './ZeroOutline';
 export { default as ZoomInOutline } from './ZoomInOutline';
 export { default as ZoomOutOutline } from './ZoomOutOutline';
-export { default as Pix } from './Pix';
-export { default as PixOutline } from './PixOutline';
-export { default as PresentationQrcodeOutline } from './PresentationQrcodeOutline';
-export { default as VariableTaxOutline } from './VariableTaxOutline';
-export { default as LockTaxOutline } from './LockTaxOutline';
-export { default as BluLogo } from './BluLogo';
